% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark_methods}
\alias{benchmark_methods}
\title{Published SL Method Benchmark (CV3)}
\format{
A data.frame with 10 rows and 3 columns:
\describe{
  \item{Method}{Method name}
  \item{CV3_AUROC}{AUROC in the most stringent CV3 setting}
  \item{Interpretability}{Whether the method is interpretable}
}
}
\usage{
benchmark_methods
}
\description{
CV3 (gene-pair isolation) AUROC values from Feng et al. (2024)
Nature Communications benchmark of 8 SL prediction methods,
compared with our DD approach.
}
\keyword{datasets}
