% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{build_mutation_matrix}
\alias{build_mutation_matrix}
\title{Build binary mutation matrix}
\usage{
build_mutation_matrix(mutations_df, cell_lines, genes)
}
\arguments{
\item{mutations_df}{A data.frame with columns DepMap_ID and Gene}

\item{cell_lines}{Character vector of cell line IDs}

\item{genes}{Character vector of gene names}
}
\value{
A binary matrix (cell lines x genes): 1 = mutated, 0 = wild-type
}
\description{
Build binary mutation matrix
}
