% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{compute_pcs}
\alias{compute_pcs}
\title{Compute Paralog Compensation Score}
\usage{
compute_pcs(
  dependency,
  expression,
  driver_gene,
  paralog_gene,
  mut_lines,
  wt_lines
)
}
\arguments{
\item{dependency}{Numeric matrix of Chronos gene-effect scores}

\item{expression}{Numeric matrix of expression values}

\item{driver_gene}{Character, the driver gene}

\item{paralog_gene}{Character, the paralog gene}

\item{mut_lines}{Mutant cell line IDs}

\item{wt_lines}{Wildtype cell line IDs}
}
\value{
A list with PCS, delta_expression, necessity
}
\description{
PCS = delta_expression(paralog, MUT vs WT) * necessity(paralog)
where necessity = -mean(gene_effect_paralog) globally.
}
