% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{filter_cancer_cell_lines}
\alias{filter_cancer_cell_lines}
\title{Filter cell lines to specific cancer types}
\usage{
filter_cancer_cell_lines(models_df, cancer_types)
}
\arguments{
\item{models_df}{Cell line annotations data.frame}

\item{cancer_types}{Character vector of Oncotree disease names to match}
}
\value{
A character vector of matching cell line IDs
}
\description{
Filter cell lines to specific cancer types
}
