% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{load_expression}
\alias{load_expression}
\title{Load DepMap gene expression data}
\usage{
load_expression(path)
}
\arguments{
\item{path}{Path to OmicsExpressionProteinCodingGenesTPMLogp1.csv}
}
\value{
A numeric matrix (cell lines x genes)
}
\description{
Load DepMap gene expression data
}
