% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{load_models}
\alias{load_models}
\title{Load DepMap cell line annotations}
\usage{
load_models(path)
}
\arguments{
\item{path}{Path to Model.csv}
}
\value{
A data.frame with cell line annotations
}
\description{
Load DepMap cell line annotations
}
