% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\docType{package}
\name{paralogSL}
\alias{paralogSL-package}
\alias{paralogSL}
\title{paralogSL: Paralog-Based Synthetic Lethality Prediction via Delta Dependency}
\description{
A lightweight R package for de novo synthetic lethality discovery through
paralog compensation analysis using DepMap CRISPR dependency data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjogzt/paralogSL}
  \item Report bugs at \url{https://github.com/tjogzt/paralogSL/issues}
}

}
\author{
\strong{Maintainer}: Tao Zhu \email{zhutao@tjh.tjmu.edu.cn} (\href{https://orcid.org/0009-0001-0779-2245}{ORCID})

Authors:
\itemize{
  \item Qingqing Mo
}

}
