% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{plot_benchmark}
\alias{plot_benchmark}
\title{Plot benchmark comparison}
\usage{
plot_benchmark(benchmark_df = get_benchmark_table())
}
\arguments{
\item{benchmark_df}{Result from get_benchmark_table()}
}
\value{
A ggplot2 object
}
\description{
Plot benchmark comparison
}
