% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{plot_protein_correlation}
\alias{plot_protein_correlation}
\title{Plot paralog co-variation from CPTAC proteomics data}
\usage{
plot_protein_correlation(prot_df, gene_a, gene_b)
}
\arguments{
\item{prot_df}{Protein abundance data.frame (genes x samples)}

\item{gene_a}{First gene name}

\item{gene_b}{Second gene name}
}
\value{
A ggplot2 object
}
\description{
Plot paralog co-variation from CPTAC proteomics data
}
