% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{stratify_by_mutation_type}
\alias{stratify_by_mutation_type}
\title{Stratify mutations by consequence type}
\usage{
stratify_by_mutation_type(mutations_df, cell_lines, driver_gene)
}
\arguments{
\item{mutations_df}{Mutation data.frame with columns DepMap_ID, Gene, VariantInfo}

\item{cell_lines}{Character vector of cell line IDs}

\item{driver_gene}{Driver gene name}
}
\value{
A list with trunc_lines and miss_lines (character vectors of cell line IDs)
}
\description{
Classifies mutations as truncating (frameshift, nonsense, splice-site)
or missense. Returns a binary matrix per mutation type.
}
