% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogSL.R
\name{compute_auroc}
\alias{compute_auroc}
\title{Compute AUROC for DD prediction against known SL pairs}
\usage{
compute_auroc(results_df, known_pairs)
}
\arguments{
\item{results_df}{Results from run_paralog_analysis()}

\item{known_pairs}{data.frame with columns gene_A, gene_B}
}
\value{
The AUROC value
}
\description{
Compute AUROC for DD prediction against known SL pairs
}
