% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cross_cancer_summary}
\alias{cross_cancer_summary}
\title{Cross-Cancer Summary}
\format{
A data.frame with columns:
\describe{
  \item{cancer}{Cancer type name}
  \item{n_lines}{Number of cell lines}
  \item{n_pairs}{Number of paralog pairs tested}
  \item{n_known}{Number of known paralog-SL pairs}
  \item{dd_auroc}{DD AUROC against gold standard}
}
}
\usage{
cross_cancer_summary
}
\description{
A data.frame of gynecological cancer cross-cancer analysis results.
}
\keyword{datasets}
