% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solid_tumor_summary}
\alias{solid_tumor_summary}
\title{Solid Tumor Pan-Cancer Summary}
\format{
A data.frame with columns:
\describe{
  \item{cancer}{Cancer type name}
  \item{n_lines}{Number of available cell lines}
}
}
\usage{
solid_tumor_summary
}
\description{
A data.frame of DD AUROC results across 24 solid tumor types.
}
\keyword{datasets}
